fix(elaborator): bind an impl's Self between its names and their bounds
A bound on an `impl` block's own type parameter may project off the block's
target (`impl<O: Uses<Self::Item>> Run for Wrap<O>`), and the target is itself
resolved from those names (`impl<U> Maker<Container<U>> for Foo<U>`). So
`register_impl_block_params` binds the names, resolves the target to establish
`Self`, and reads the bounds last. A bound read before `Self` stood for the
target was pinned to no receiver and rejected outright.

Distilled alongside: `with_self_binding` saves, installs and restores one
`SelfFrame`, so the three facts making up `Self` are listed in one place;
`under_self_binding` replaces three copies of the optional-binding match; the
crate rule on `assert!` says what an assertion is for rather than leaving the
question open; and the spec and WEP record which declarations bind a `Self`
their bounds may project off, with the data declarations that do not as a gap.

+++ wado-compiler/src/elaborator/orchestration.rs
@@ -3626,7 +3626,7 @@ impl<'a, H: CompilerHost> Elaborator<'a, H> {
                 // arguments, so it is where the target names the parameter.
                 let param_at = |name: &str| {
                     let param = param_slots.iter().find(|p| p.name == name)?;
-                    Some((target_arg_slot(&impl_block.ty, name)? as usize, param))
+                    Some((target_arg_slot(&impl_block.ty, name)?, param))
                 };
 
                 for binding in &impl_block.associated_types {
@@ -3638,7 +3638,7 @@ impl<'a, H: CompilerHost> Elaborator<'a, H> {
                             };
                             type_table
                                 .borrow_mut()
-                                .make_type_param(named.name.clone(), idx as u32)
+                                .make_type_param(named.name.clone(), idx)
                         }
                         // Chained case: `type Item = I::InnerName` — I is a type param
                         Type::NamespacedGeneric(ns) if ns.args.is_empty() => {
@@ -3654,7 +3654,7 @@ impl<'a, H: CompilerHost> Elaborator<'a, H> {
                             };
                             let inner_param_id = type_table
                                 .borrow_mut()
-                                .make_type_param(ns.namespace.clone(), idx as u32);
+                                .make_type_param(ns.namespace.clone(), idx);
                             type_table.borrow_mut().make_assoc_type_projection(
                                 inner_param_id,
                                 owning_trait,
